dvg/backend/requirements.txt
asgiref==3.7.2
Django==3.2.22
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.7.1
graphene-django==2.14.0
django-cors-headers==3.6.0
django-ckeditor==6.3.2
Pillow==9.5.0
